Divido el módulo es_query
