+++ series_tiempo_ar_api/apps/api/query/es_query/__init__.py
@@ -1,0 +1,1 @@
+#! coding: utf-8
