+++ series_tiempo_ar_api/apps/api/query/es_query/series.py
@@ -1,0 +1,19 @@
+#! coding: utf-8
+from elasticsearch_dsl import Search
+
+
+class Series(object):
+    def __init__(self, series_id=None, rep_mode=None, search=None):
+        self.series_id = series_id
+        self.rep_mode = rep_mode
+        self.search = search or Search()
+        self.meta = None
+
+    def __getitem__(self, item):
+        return self.__getattribute__(item)
+
+    def __setitem__(self, key, value):
+        return self.__setattr__(key, value)
+
+    def get(self, item, default=None):
+        return getattr(self, item, default)
